+++ python/fate_test/pyproject.toml
@@ -33,7 +33,7 @@ loguru = "^0.5.1"
 prettytable = "^1.0.0"
 sshtunnel = "^0.1.5"
 fate_test = "^0.1.0"
-pandas = "^1.1.0"
+pandas = "^0.23.4"
 
 [tool.poetry.dev-dependencies]
